requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
python-dotenv==0.21.1
supabase==0.7.1
pydantic==1.10.2 